Build new class and import functions from main to it as a method
