+++ main.rb
@@ -1,61 +1,8 @@
-require_relative 'app'
-
-def app_intro
-  puts 'Welcome to the School Library App!'
-  puts ''
-  loop do
-    display_options
-    option = take_input
-    option_methods(option)
-
-    break if option == 7
-  end
-end
-
-def display_options
-  puts 'Please choose an option by enterin a number: '
-  puts '1 - List all books'
-  puts '2 - List all people'
-  puts '3 - Create a person'
-  puts '4 - Create a book'
-  puts '5 - Create a rental'
-  puts '6 - List all rentals for a given person ID'
-  puts '7 - Exit'
-end
-
-def take_input
-  option = gets.chomp.to_i
-  until option >= 1 || option <= 7
-    puts 'Please enter a number between 1 and 7!'
-    print 'Enter a number: '
-    option = gets.chomp.to_i
-  end
-  option
-end
-
-def option_methods(option)
-  case option
-  when 1
-    list_all_books
-  when 2
-    list_all_persons
-  when 3
-    create_person
-  when 4
-    create_book
-  when 5
-    create_rental
-  when 6
-    list_all_rentals_of_person
-  else
-    puts 'Thank you for using this app!'
-    exit
-  end
-end
+require_relative 'app_controller'
 
 def main
-  app = App.new
-  app.run
+  # app = App.new
+  # app.run
 end
 
 main
